Keep VPC and ACL in line with existing footer; break tags out
