+++ packages/manager/src/features/Kubernetes/KubernetesClusterDetail/KubeEntityDetailFooter.tsx
@@ -1,27 +1,36 @@
-import { useProfile } from '@linode/queries';
+import { useProfile, useVPCQuery } from '@linode/queries';
+import { Box, CircleProgress, StyledLinkButton } from '@linode/ui';
+import { pluralize } from '@linode/utilities';
 import Grid from '@mui/material/Grid';
 import { useTheme } from '@mui/material/styles';
-import { useSnackbar } from 'notistack';
 import * as React from 'react';
 
-import { TagCell } from 'src/components/TagCell/TagCell';
+import { Link } from 'src/components/Link';
 import {
-  StyledBox,
   StyledLabelBox,
   StyledListItem,
   sxLastListItem,
   sxListItemFirstChild,
 } from 'src/features/Linodes/LinodeEntityDetail.styles';
-import { useKubernetesClusterMutation } from 'src/queries/kubernetes';
-import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
 import { formatDate } from 'src/utilities/formatDate';
+
+import { useIsLkeEnterpriseEnabled } from '../kubeUtils';
+
+import type { KubernetesControlPlaneACLPayload } from '@linode/api-v4';
+import type { SxProps } from '@mui/material/styles';
 interface FooterProps {
+  aclData: KubernetesControlPlaneACLPayload | undefined;
   areClusterLinodesReadOnly: boolean;
   clusterCreated: string;
   clusterId: number;
   clusterLabel: string;
   clusterTags: string[];
   clusterUpdated: string;
+  isClusterReadOnly: boolean;
+  isLoadingKubernetesACL: boolean;
+  setControlPlaneACLDrawerOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  sx?: SxProps;
+  vpcId: number | undefined;
 }
 
 export const KubeEntityDetailFooter = React.memo((props: FooterProps) => {
@@ -29,33 +38,30 @@ export const KubeEntityDetailFooter = React.memo((props: FooterProps) => {
 
   const { data: profile } = useProfile();
   const {
-    areClusterLinodesReadOnly,
+    aclData,
     clusterCreated,
     clusterId,
-    clusterLabel,
-    clusterTags,
     clusterUpdated,
+    isLoadingKubernetesACL,
+    isClusterReadOnly,
+    setControlPlaneACLDrawerOpen,
+    vpcId,
   } = props;
 
-  const { mutateAsync: updateKubernetesCluster } =
-    useKubernetesClusterMutation(clusterId);
+  const { isLkeEnterprisePhase2FeatureEnabled } = useIsLkeEnterpriseEnabled();
 
-  const { enqueueSnackbar } = useSnackbar();
+  const enabledACL = aclData?.acl.enabled ?? false;
+  const totalIPv4 = aclData?.acl.addresses?.ipv4?.length ?? 0;
+  const totalIPv6 = aclData?.acl.addresses?.ipv6?.length ?? 0;
+  const totalNumberIPs = totalIPv4 + totalIPv6;
 
-  const handleUpdateTags = React.useCallback(
-    (newTags: string[]) => {
-      return updateKubernetesCluster({
-        tags: newTags,
-      }).catch((e) =>
-        enqueueSnackbar(
-          getAPIErrorOrDefault(e, 'Error updating tags')[0].reason,
-          {
-            variant: 'error',
-          }
-        )
-      );
-    },
-    [updateKubernetesCluster, enqueueSnackbar]
+  const buttonCopyACL = enabledACL
+    ? `Enabled (${pluralize('IP Address', 'IP Addresses', totalNumberIPs)})`
+    : 'Enable';
+
+  const { data: vpc } = useVPCQuery(
+    vpcId ?? -1,
+    isLkeEnterprisePhase2FeatureEnabled
   );
 
   return (
@@ -83,19 +89,66 @@ export const KubeEntityDetailFooter = React.memo((props: FooterProps) => {
 
           [theme.breakpoints.down('lg')]: {
             padding: '8px',
-            paddingTop: 0,
           },
 
-          [theme.breakpoints.down('md')]: {
+          [theme.breakpoints.between('md', 'xs')]: {
             display: 'grid',
             gridTemplateColumns: '50% 2fr',
           },
         }}
       >
-        <StyledBox>
+        <Box
+          sx={{
+            display: 'flex',
+            flexWrap: 'wrap',
+            rowGap: 1,
+          }}
+        >
           <StyledListItem sx={{ ...sxListItemFirstChild }}>
             <StyledLabelBox component="span">Cluster ID:</StyledLabelBox>{' '}
             {clusterId}
+          </StyledListItem>
+          {isLkeEnterprisePhase2FeatureEnabled && vpc && (
+            <StyledListItem
+              sx={{
+                alignItems: 'center',
+              }}
+            >
+              <StyledLabelBox component="span">VPC: </StyledLabelBox>{' '}
+              <Link
+                data-testid="assigned-lke-cluster-label"
+                to={`/vpcs/${vpcId}`}
+              >
+                {vpc?.label ?? `${vpcId}`}
+              </Link>
+              &nbsp;
+              {vpcId && vpc?.label ? `(ID: ${vpcId})` : undefined}
+            </StyledListItem>
+          )}
+          <StyledListItem>
+            <StyledLabelBox component="span">
+              Control Plane ACL:{' '}
+            </StyledLabelBox>{' '}
+            {isLoadingKubernetesACL ? (
+              <Box component="span" sx={{ paddingLeft: 1 }}>
+                <CircleProgress noPadding size="sm" />
+              </Box>
+            ) : (
+              <StyledLinkButton
+                disabled={isClusterReadOnly}
+                onClick={() => setControlPlaneACLDrawerOpen(true)}
+                sx={(theme) => ({
+                  '&:disabled': {
+                    '& g': {
+                      stroke: theme.tokens.alias.Content.Icon.Primary.Disabled,
+                    },
+                    color: theme.tokens.alias.Content.Text.Primary.Disabled,
+                  },
+                })}
+              >
+                {buttonCopyACL}
+              </StyledLinkButton>
+            )}
           </StyledListItem>
           <StyledListItem>
             <StyledLabelBox component="span">Created:</StyledLabelBox>{' '}
@@ -109,32 +162,7 @@ export const KubeEntityDetailFooter = React.memo((props: FooterProps) => {
               timezone: profile?.timezone,
             })}
           </StyledListItem>
-        </StyledBox>
-      </Grid>
-      <Grid
-        size={{
-          lg: 3.5,
-          xs: 12,
-        }}
-        sx={{
-          marginLeft: 'auto',
-
-          [theme.breakpoints.down('lg')]: {
-            display: 'flex',
-            justifyContent: 'flex-start',
-          },
-        }}
-      >
-        <TagCell
-          disabled={areClusterLinodesReadOnly}
-          entityLabel={clusterLabel}
-          sx={{
-            width: '100%',
-          }}
-          tags={clusterTags}
-          updateTags={handleUpdateTags}
-          view="inline"
-        />
+        </Box>
       </Grid>
     </Grid>
   );
