+++ packages/manager/src/features/Kubernetes/KubernetesClusterDetail/KubeSummaryPanel.tsx
@@ -15,19 +15,23 @@ import { ActionMenu } from 'src/components/ActionMenu/ActionMenu';
 import { ConfirmationDialog } from 'src/components/ConfirmationDialog/ConfirmationDialog';
 import { EntityDetail } from 'src/components/EntityDetail/EntityDetail';
 import { EntityHeader } from 'src/components/EntityHeader/EntityHeader';
+import { TagCell } from 'src/components/TagCell/TagCell';
 import { KubeClusterSpecs } from 'src/features/Kubernetes/KubernetesClusterDetail/KubeClusterSpecs';
 import { useIsLkeEnterpriseEnabled } from 'src/features/Kubernetes/kubeUtils';
 import { useIsResourceRestricted } from 'src/hooks/useIsResourceRestricted';
 import {
+  useKubernetesClusterMutation,
   useKubernetesControlPlaneACLQuery,
   useKubernetesDashboardQuery,
   useResetKubeConfigMutation,
 } from 'src/queries/kubernetes';
-import { getErrorStringOrDefault } from 'src/utilities/errorUtils';
+import {
+  getAPIErrorOrDefault,
+  getErrorStringOrDefault,
+} from 'src/utilities/errorUtils';
 
 import { ClusterChips } from '../ClusterList/ClusterChips';
 import { DeleteKubernetesClusterDialog } from './DeleteKubernetesClusterDialog';
-import { KubeAddOnsFooter } from './KubeAddOnsFooter';
 import { KubeConfigDisplay } from './KubeConfigDisplay';
 import { KubeConfigDrawer } from './KubeConfigDrawer';
 import { KubeControlPlaneACLDrawer } from './KubeControlPaneACLDrawer';
@@ -81,6 +85,26 @@ export const KubeSummaryPanel = React.memo((props: Props) => {
 
   const [resetKubeConfigDialogOpen, setResetKubeConfigDialogOpen] =
     React.useState(false);
+
+  const { mutateAsync: updateKubernetesCluster } = useKubernetesClusterMutation(
+    cluster.id
+  );
+
+  const handleUpdateTags = React.useCallback(
+    (newTags: string[]) => {
+      return updateKubernetesCluster({
+        tags: newTags,
+      }).catch((e) =>
+        enqueueSnackbar(
+          getAPIErrorOrDefault(e, 'Error updating tags')[0].reason,
+          {
+            variant: 'error',
+          }
+        )
+      );
+    },
+    [updateKubernetesCluster, enqueueSnackbar]
+  );
 
   const handleResetKubeConfig = () => {
     return resetKubeConfig({ id: cluster.id }).then(() => {
@@ -141,26 +165,43 @@ export const KubeSummaryPanel = React.memo((props: Props) => {
             sx={{
               flex: 1,
               justifyContent: 'space-between',
+              paddingTop: '8px',
             }}
           >
-            <KubeAddOnsFooter
-              aclData={aclData}
-              isClusterReadOnly={isClusterReadOnly}
-              isLoadingKubernetesACL={isLoadingKubernetesACL}
-              setControlPlaneACLDrawerOpen={setControlPlaneACLDrawerOpen}
-              sx={{
-                paddingTop: 2,
-              }}
-              vpcId={cluster.vpc_id}
-            />
             <KubeEntityDetailFooter
+              aclData={aclData}
               areClusterLinodesReadOnly={areClusterLinodesReadOnly}
               clusterCreated={cluster.created}
               clusterId={cluster.id}
               clusterLabel={cluster.label}
               clusterTags={cluster.tags}
               clusterUpdated={cluster.updated}
+              isClusterReadOnly={isClusterReadOnly}
+              isLoadingKubernetesACL={isLoadingKubernetesACL}
+              setControlPlaneACLDrawerOpen={setControlPlaneACLDrawerOpen}
+              vpcId={cluster.vpc_id}
             />
+            <Grid
+              size={{
+                lg: 12,
+                xs: 12,
+              }}
+              sx={{
+                marginLeft: 'auto',
+                marginTop: '8px',
+              }}
+            >
+              <TagCell
+                disabled={areClusterLinodesReadOnly}
+                entityLabel={cluster.label}
+                sx={{
+                  width: '100%',
+                }}
+                tags={cluster.tags}
+                updateTags={handleUpdateTags}
+                view="inline"
+              />
+            </Grid>
           </Grid>
         }
         header={
